+++ src/provider/aws/elasticache/cmd/transitEncryptionEnabled.ts
@@ -3,7 +3,7 @@ import assert from 'assert'
 import { AWS, keyTypeArg } from '../../../../lib/aws/AWS'
 import { CacheCluster } from 'aws-sdk/clients/elasticache'
 
-const rule = 'EncryptionAtRest'
+const rule = 'TransitEncryptionEnabled'
 
 export const builder = {
   ...keyTypeArg,
@@ -11,12 +11,11 @@ export const builder = {
 
 export const command = `${rule} [args]`
 
-export const desc = `Elasticache instances should have storage encrypted at rest
+export const desc = `Elasticache instances should enforce encryption in flight
 
-  OK      - Elasticache cluster storage is encrypted at rest
-  WARNING - Elasticache cluster storage is encrypted at rest, but with the wrong key type
-  UNKNOWN - Unable to determine if the elasticache storage is encrypted
-  FAIL    - Elasticache cluster storage is not encrypted at rest
+  OK      - Elasticache cluster has transit encryption enabled
+  UNKNOWN - Unable to determine if the elasticache cluster has transit encryption
+  FAIL    - Elasticache cluster does not have encryption enabled
 
   resourceId: cluster, instance, or shard name
 
@@ -25,7 +24,7 @@ export const desc = `Elasticache instances should have storage encrypted at rest
 
 `
 
-export default class AtRestEncryption extends AWS {
+export default class TransitEncryptionEnabled extends AWS {
   audits: AuditResultInterface[] = []
   service = 'elasticache'
   global = false
@@ -60,68 +59,13 @@ export default class AtRestEncryption extends AWS {
       time: new Date().toISOString(),
     }
 
-    if (resource.AtRestEncryptionEnabled) {
-      if (resource.Engine === 'redis') {
-        await this.handleRedis({
-          resource,
-          region,
-          audit,
-        })
-      }
+    if (resource.TransitEncryptionEnabled) {
+      audit.state = 'OK'
     } else {
       audit.state = 'FAIL'
     }
 
     this.audits.push(audit)
-  }
-
-  handleRedis = async ({
-    resource,
-    region,
-    audit,
-  }: {
-    resource: CacheCluster
-    region: string
-    audit: AuditResultInterface
-  }) => {
-    const options = this.getOptions()
-    options.region = region
-
-    assert(
-      resource.ReplicationGroupId,
-      'all redis clusters must have a replication group id'
-    )
-
-    const groups = await new this.AWS.ElastiCache(options)
-      .describeReplicationGroups({
-        ReplicationGroupId: resource.ReplicationGroupId,
-      })
-      .promise()
-
-    if (groups.ReplicationGroups) {
-      assert(
-        Array.isArray(groups.ReplicationGroups),
-        'expecting group to be an array'
-      )
-
-      assert(
-        groups.ReplicationGroups.length === 1,
-        'expecting a single resource in the group'
-      )
-
-      for (const group of groups.ReplicationGroups) {
-        if (group.KmsKeyId) {
-          assert(this.keyType, 'key type is required')
-          audit.state = await this.isKeyTrusted(
-            group.KmsKeyId,
-            this.keyType,
-            region
-          )
-        } else {
-          audit.state = 'FAIL'
-        }
-      }
-    }
   }
 
   scan = async ({
@@ -148,7 +92,7 @@ export default class AtRestEncryption extends AWS {
 }
 
 export const handler = async (args: AWSScannerInterface) => {
-  const scanner = new AtRestEncryption({
+  const scanner = new TransitEncryptionEnabled({
     region: args.region,
     profile: args.profile,
     resourceId: args.resourceId,