+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- AWS - elasticache - TransitEncryptionEnabled
+
 ## [1.4.0] - 2020-10-19
 
 ### Added